Repair failing tests.
