+++ integrations/datastore/src/main/java/de/kaiserpfalzedv/rpg/integrations/datastore/file/store/FileCouldNotBeSavedException.java
@@ -15,25 +15,23 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.kaiserpfalzedv.rpg.integrations.datastore;
-
-import de.kaiserpfalzedv.rpg.integrations.datastore.store.FileHandlingException;
+package de.kaiserpfalzedv.rpg.integrations.datastore.file.store;
 
 import java.util.UUID;
 
 /**
- * DuplicateFileException -- There is a file stored under this data.
+ * FileCouldNotBeSavedException -- The file could not be saved.
  *
  * @author klenkes74 {@literal <rlichti@kaiserpfalz-edv.de>}
  * @since 1.0.0 2021-01-08
  */
-public class DuplicateFileException extends FileHandlingException {
+public class FileCouldNotBeSavedException extends FileHandlingException {
     /**
      * @param uid UID of the file resource.
-     * @param message The failure message.
+     * @param message the failure message.
      */
     @SuppressWarnings("CdiInjectionPointsInspection")
-    public DuplicateFileException(final UUID uid, final String message) {
+    public FileCouldNotBeSavedException(final UUID uid, final String message) {
         super(uid, message);
     }
 
@@ -41,7 +39,7 @@ public class DuplicateFileException extends FileHandlingException {
      * @param uid UID of the file resource.
      * @param cause The failure cause.
      */
-    public DuplicateFileException(final UUID uid, final Throwable cause) {
+    public FileCouldNotBeSavedException(final UUID uid, final Throwable cause) {
         super(uid, cause);
     }
 
@@ -50,7 +48,7 @@ public class DuplicateFileException extends FileHandlingException {
      * @param message The failure message.
      * @param cause The failure cause.
      */
-    public DuplicateFileException(final UUID uid, final String message, final Throwable cause) {
+    public FileCouldNotBeSavedException(final UUID uid, final String message, final Throwable cause) {
         super(uid, message, cause);
     }
 }
