+++ integrations/datastore/src/main/java/de/kaiserpfalzedv/rpg/integrations/datastore/file/store/FileResourceService.java
@@ -15,9 +15,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.kaiserpfalzedv.rpg.integrations.datastore.store;
+package de.kaiserpfalzedv.rpg.integrations.datastore.file.store;
 
-import de.kaiserpfalzedv.rpg.integrations.datastore.DuplicateFileException;
+import de.kaiserpfalzedv.rpg.integrations.datastore.file.DuplicateFileException;
 
 import java.io.InputStream;
 import java.util.UUID;
