+++ integrations/datastore/src/test/java/de/kaiserpfalzedv/rpg/integrations/datastore/file/TestResource.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.kaiserpfalzedv.rpg.core.resources;
+package de.kaiserpfalzedv.rpg.integrations.datastore.file;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -24,10 +24,12 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.util.StdDateFormat;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import de.kaiserpfalzedv.rpg.core.dice.TestDice;
-import de.kaiserpfalzedv.rpg.core.files.FileResource;
-import de.kaiserpfalzedv.rpg.core.files.ImmutableFileData;
-import de.kaiserpfalzedv.rpg.core.files.ImmutableFileResource;
+import de.kaiserpfalzedv.rpg.core.resources.ImmutableResourceHistory;
+import de.kaiserpfalzedv.rpg.core.resources.ImmutableResourceMetadata;
+import de.kaiserpfalzedv.rpg.core.resources.ImmutableResourceStatus;
+import de.kaiserpfalzedv.rpg.integrations.datastore.file.store.FileResource;
+import de.kaiserpfalzedv.rpg.integrations.datastore.file.store.ImmutableFileData;
+import de.kaiserpfalzedv.rpg.integrations.datastore.file.store.ImmutableFileResource;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,7 +45,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class TestResource {
     static private final Logger LOG = LoggerFactory.getLogger("TestResource");
 
-    public static final String JSON_STRING = "{\"metadata\":{\"uid\":\"7ee3f57b-ed6c-400e-84f4-253afe0c56c0\",\"generation\":1,\"created\":\"2021-01-07T18:10:20.168433Z\",\"annotations\":{},\"labels\":{},\"kind\":\"File\",\"apiVersion\":\"v1\",\"namespace\":\"test\",\"name\":\"TestFile\",\"selfLink\":\"/apis/files/v1/namespaces/test/File/7ee3f57b-ed6c-400e-84f4-253afe0c56c0\"},\"spec\":{\"size\":0,\"mimeType\":\"text/plain\"},\"status\":{\"observedGeneration\":1,\"history\":[{\"timeStamp\":\"2021-01-07T18:10:20.168433Z\",\"status\":\"created\",\"message\":\"File saved.\"}]}}";
+    public static final String JSON_STRING = "{\"metadata\":{\"kind\":\"File\",\"apiVersion\":\"v1\",\"namespace\":\"test\",\"name\":\"TestFile\",\"selfLink\":\"/apis/files/v1/namespaces/test/File/7ee3f57b-ed6c-400e-84f4-253afe0c56c0\",\"uid\":\"7ee3f57b-ed6c-400e-84f4-253afe0c56c0\",\"generation\":1,\"created\":\"2021-01-07T18:10:20.168433Z\",\"annotations\":{},\"labels\":{}},\"spec\":{\"size\":0,\"mimeType\":\"text/plain\"},\"status\":{\"observedGeneration\":1,\"history\":[{\"timeStamp\":\"2021-01-07T18:10:20.168433Z\",\"status\":\"created\",\"message\":\"File saved.\"}]}}";
     private static final UUID UID = UUID.fromString("7ee3f57b-ed6c-400e-84f4-253afe0c56c0");
     private static final OffsetDateTime CREATED = OffsetDateTime.parse("2021-01-07T18:10:20.168433Z");
 
@@ -119,7 +121,7 @@ public class TestResource {
 
     @BeforeAll
     static void setUp() {
-        MDC.put("test-class", TestDice.class.getSimpleName());
+        MDC.put("test-class", TestResource.class.getSimpleName());
     }
 
     @AfterAll
