+++ integrations/drivethru/pom.xml
@@ -53,47 +53,4 @@
             <artifactId>quarkus-cache</artifactId>
         </dependency>
     </dependencies>
-
-    <build>
-        <resources>
-            <resource>
-                <filtering>true</filtering>
-                <directory>src/main/resources</directory>
-                <excludes>
-                    <exclude>**/*.docx</exclude>
-                    <exclude>**/*.xlsx</exclude>
-                    <exclude>**/*.zip</exclude>
-                    <exclude>**/*.tgz</exclude>
-                    <exclude>**/*.png</exclude>
-                    <exclude>**/*.jpg</exclude>
-                    <exclude>**/*.gpg</exclude>
-                    <exclude>**/*.license</exclude>
-                    <exclude>**/*.js</exclude>
-                    <exclude>**/*.js.*</exclude>
-                    <exclude>**/*.ttf</exclude>
-                    <exclude>**/*.woff*</exclude>
-                    <exclude>**/*.eot</exclude>
-                </excludes>
-            </resource>
-            <resource>
-                <filtering>false</filtering>
-                <directory>src/main/resources</directory>
-                <includes>
-                    <include>**/*.docx</include>
-                    <include>**/*.xlsx</include>
-                    <include>**/*.zip</include>
-                    <include>**/*.tgz</include>
-                    <include>**/*.png</include>
-                    <include>**/*.jpg</include>
-                    <include>**/*.gpg</include>
-                    <include>**/*.license</include>
-                    <include>**/*.js</include>
-                    <include>**/*.js.*</include>
-                    <include>**/*.ttf</include>
-                    <include>**/*.woff*</include>
-                    <include>**/*.eot</include>
-                </includes>
-            </resource>
-        </resources>
-    </build>
 </project>
