+++ rpg-core/src/test/java/de/kaiserpfalzedv/rpg/core/snowflake/SnowflakeTest.java
@@ -16,25 +16,17 @@
 
 package de.kaiserpfalzedv.rpg.core.snowflake;
 
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * @author downgoon {@literal http://downgoon.com}
  * @since 1.0.0 2021-01-11
  */
 public class SnowflakeTest {
-
-	@BeforeClass
-	public static void setUpBeforeClass() throws Exception {
-	}
-
-	@Before
-	public void setUp() throws Exception {
-	}
-
 	@Test
 	public void testNextIdAndParse() throws Exception {
 		long beginTimeStamp = System.currentTimeMillis();
@@ -45,21 +37,21 @@ public class SnowflakeTest {
 		long[] arr = snowflake.parseId(id);
 		System.out.println(snowflake.formatId(id));
 		
-		Assert.assertTrue(arr[0] >= beginTimeStamp);
-		Assert.assertEquals(3, arr[1]); // datacenterId
-		Assert.assertEquals(16, arr[2]); // workerId
-		Assert.assertEquals(0, arr[3]); // sequenceId
+		assertTrue(arr[0] >= beginTimeStamp);
+		assertEquals(3, arr[1]); // datacenterId
+		assertEquals(16, arr[2]); // workerId
+		assertEquals(0, arr[3]); // sequenceId
 
 		// gen two ids in different MS
 		long id2 = snowflake.nextId();
-		Assert.assertFalse(id == id2);
+		assertNotEquals(id2, id);
 		System.out.println(snowflake.formatId(id2));
 		 
 		Thread.sleep(1); // wait one ms 
 		long id3 = snowflake.nextId();
 		long[] arr3 = snowflake.parseId(id3);
 		System.out.println(snowflake.formatId(id3));
-		Assert.assertTrue(arr3[0] > arr[0]);
+		assertTrue(arr3[0] > arr[0]);
 		
 		// gen two ids in the same MS
 		long[] ids = new long[2];
@@ -68,11 +60,11 @@ public class SnowflakeTest {
 		}
 		System.out.println(snowflake.formatId(ids[0]));
 		System.out.println(snowflake.formatId(ids[1]));
-		Assert.assertFalse(ids[0] == ids[1]);
+		assertNotEquals(ids[1], ids[0]);
 		long[] arr_ids0 = snowflake.parseId(ids[0]);
 		long[] arr_ids1 = snowflake.parseId(ids[1]);
-		Assert.assertEquals(arr_ids0[0], arr_ids1[0]);
-		Assert.assertEquals(arr_ids0[3], arr_ids1[3] - 1);
+		assertEquals(arr_ids0[0], arr_ids1[0]);
+		assertEquals(arr_ids0[3], arr_ids1[3] - 1);
 	}
 	
 }
