+++ rpg-server/tomb/pom.xml
@@ -70,6 +70,7 @@
         <dependency>
             <groupId>net.dv8tion</groupId>
             <artifactId>JDA</artifactId>
+            <version>4.2.0_227</version>
         </dependency>
 
         <dependency>
